Add `envar` writer to generate code for `jsr:@wuespace/envar`
With this change, it is possible to, based on the example env file,
automatically generate code to use with the `jsr:@wuespace/envar` package.

This generates a self-contained file which exports a function `initEnv(): Promise<void>`,
which initializes the variables with basic validators.

Usage:

```shell
envardoc envar -o lib/env.ts example.env
```