"Stateless" release automation
With this, releases (and publishing) are triggered automatically based
on the push of a tag (e.g., by creating a release in the GitHub UI).

Version-specific strings get replaced automatically, so a release doesn't
require any additional commits. With this, releases can, essentially, be
"put" (as a tag) on any commit, without having to worry about updating version
numbers everywhere.